manage/requirements.txt
uvicorn[standard]==0.17.6
databases==0.6.2
fastapi==0.89.1
SQLAlchemy==1.4.41
sqlalchemy-json==0.5.0
pandas==1.5.1
pydantic==1.10.2
python-decouple==3.6
psycopg2-binary==2.9.5
passlib==1.7.4
python-multipart==0.0.5
awswrangler==2.18.0
boto3==1.26.24
PyJWT==2.6.0
bcrypt==4.0.1
python-jose==3.3.0
cryptography==38.0.4
starlette==0.22.0
fastapi-mail==1.2.4
sendgrid
qrcode==7.4.2
pyotp==2.8.0